Chapter 11 commit
So I can't say exactly where the code breaks but I followed the "textbook" if you can call it that (in my opinion, this thing is the physical manifestation of the Uninformative Youtube Programming tutorial cliche) but now the primary tasklist doesn't appear properly and I don't know how long it's been broken for. Debugger makes it appear very possible to be lack of an adapter but I don't even know where to start looking thanks to the aforementioned issue. I just want to get this over with before my sanity breaks and I have any more reactions to this stress.